BOAC-3809, break up AcademicTimeline component for readability
